move pkg/tenancy to internal/tenancy
